--- apps/client/src/util/bookmarkFunctions.ts
@@ -1,73 +1,0 @@
-import { useMutation, useQueryClient } from '@tanstack/react-query';
-
-import { useGuardContext } from '@/shared/context/hooks.ts';
-import { ServiceContext } from '@/shared/context/ServiceContext.ts';
-import { TokenContext } from '@/shared/context/TokenContext.ts';
-
-export const useAddBookmark = () => {
-  const { postService } = useGuardContext(ServiceContext);
-  const { token } = useGuardContext(TokenContext);
-
-  const queryClient = useQueryClient();
-
-  const { mutate: addBookmark, isPending } = useMutation({
-    mutationFn: ({ postId }: { postId: string }) => {
-      if (token === null) {
-        throw new Error('토큰이 존재하지 않습니다.');
-      }
-      return postService.addBookmark({ token, postId });
-    },
-    onSuccess: async (response) => {
-      if (response.type === 'success') {
-        await queryClient.invalidateQueries({ queryKey: ['postService'] });
-        return;
-      } else {
-        // TODO: 북마크 생성 실패 시 하단에 토스트 띄우기
-        return;
-      }
-    },
-    onError: () => {
-      // TODO: 북마크 생성 실패 시 하단에 토스트 띄우기
-      return;
-    },
-  });
-
-  return {
-    addBookmark,
-    isPending,
-  };
-};
-
-export const useDeleteBookmark = () => {
-  const { postService } = useGuardContext(ServiceContext);
-  const { token } = useGuardContext(TokenContext);
-
-  const queryClient = useQueryClient();
-
-  const { mutate: deleteBookmark, isPending } = useMutation({
-    mutationFn: ({ postId }: { postId: string }) => {
-      if (token === null) {
-        throw new Error('토큰이 존재하지 않습니다.');
-      }
-      return postService.deleteBookmark({ token, postId });
-    },
-    onSuccess: async (response) => {
-      if (response.type === 'success') {
-        await queryClient.invalidateQueries({ queryKey: ['postService'] });
-        return;
-      } else {
-        // TODO: 북마크 생성 실패 시 하단에 토스트 띄우기
-        return;
-      }
-    },
-    onError: () => {
-      // TODO: 북마크 생성 실패 시 하단에 토스트 띄우기
-      return;
-    },
-  });
-
-  return {
-    deleteBookmark,
-    isPending,
-  };
-};
